moved the cc transform to the lifter file because it is used only with a lifter.
